move common widgets into shared folder
